// x-pack/platform/packages/shared/kbn-inference-tracing/src/types.d.ts
export declare enum GenAISemanticConventions {
    GenAIUsageCost = "gen_ai.usage.cost",
    GenAIUsageInputTokens = "gen_ai.usage.input_tokens",
    GenAIUsageCachedInputTokens = "gen_ai.usage.cached_input_tokens",
    GenAIUsageOutputTokens = "gen_ai.usage.output_tokens",
    GenAIOperationName = "gen_ai.operation.name",
    GenAIResponseModel = "gen_ai.response.model",
    GenAIRequestModel = "gen_ai.request.model",
    GenAISystem = "gen_ai.system",
    GenAIProviderName = "gen_ai.provider.name",
    GenAIOutputType = "gen_ai.output.type",
    GenAIToolCallId = "gen_ai.tool.call.id",
    GenAIToolName = "gen_ai.tool.name",
    GenAISystemMessage = "gen_ai.system.message",
    GenAIUserMessage = "gen_ai.user.message",
    GenAIAssistantMessage = "gen_ai.assistant.message",
    GenAIToolMessage = "gen_ai.tool.message",
    GenAIChoice = "gen_ai.choice",
    GenAIAgentId = "gen_ai.agent.id",
    GenAIConversationId = "gen_ai.conversation.id"
}